refacto config
